requirements.txt
absl-py==1.0.0
alabaster==0.7.12
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.5
asyncio==3.4.3
attrdict==2.0.1
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bokeh==2.4.3
boto3==1.23.5
botocore==1.26.5
bravado==11.0.3
bravado-core==5.17.0
Brotli==1.0.9
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
docutils==0.17.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.0
fonttools==4.33.3
fqdn==1.5.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
h2o==3.36.1.1
html5lib==1.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
jsonpointer==2.3
jsonref==0.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
kneed==0.7.0
lazy_loader==0.1rc2
lib==4.0.0
llvmlite==0.38.1
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
msgpack==1.0.3
multipledispatch==0.6.0
natsort==8.1.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
neptune-client==0.16.2
neptune-contrib==0.28.1
nest-asyncio==1.5.5
networkx==2.8.2
nose==1.3.7
notebook==6.4.11
numba==0.55.1
numpy<=1.21
numpydoc==1.3.1
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-flavor==0.3.0
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaml==21.10.1
pycparser==2.21
Pygments==2.12.0
pyjanitor==0.23.1
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
QtPy==2.1.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
s3transfer==0.5.2
scikit-learn==1.1.1
scikit-opt==0.6.6
scikit-optimize==0.9.0
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.40.0
simplejson==3.17.6
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
SQLAlchemy==1.4.36
stack-data==0.2.0
statsmodels==0.13.2
swagger-spec-validator==2.7.4
tables==3.7.0
tabulate==0.8.9
terminado==0.15.0
testpath==0.6.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typed-ast==1.5.4
typing_extensions==4.2.0
uri-template==1.2.0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
xarray==2022.3.0
zipp==3.8.0
